offsite/requirements.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.5
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
Django==3.0.7
django-celery-beat==2.0.0
django-cors-headers==3.4.0
django-rest-knox==4.1.0
django-timezone-field==4.0
djangorestframework==3.11.0
future==0.18.2
h11==0.9.0
httptools==0.1.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
kombu==4.6.10
loguru==0.5.1
pathspec==0.8.0
pkg-resources==0.0.0
psutil==5.7.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
redis==3.5.3
regex==2020.6.8
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
uWSGI==2.0.19.1
validators==0.15.0
vine==1.3.0
websockets==8.1
zope.interface==5.1.0
